% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shelve.R
\name{delete}
\alias{delete}
\title{Delete objects from active shelf}
\usage{
delete(x, shelf = shelf_active(), ignore_not_found = TRUE)
}
\arguments{
\item{x}{List of object names}

\item{shelf}{Name of shelf}

\item{ignore_not_found}{Ignore objects which do not exist.}
}
\description{
Delete objects from active shelf
}
