% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shelve.R
\name{shelf_activate}
\alias{shelf_activate}
\title{Activate shelf}
\usage{
shelf_activate(name)
}
\arguments{
\item{name}{Name of shelf. Creates one with the name if it does not exist.}
}
\description{
Shelf is created if it does not exist
}
