% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shelve.R
\name{shelf_clear}
\alias{shelf_clear}
\title{Delete all objects in shelf}
\usage{
shelf_clear(name = shelf_active())
}
\arguments{
\item{name}{Name of shelf.}
}
\description{
Deleta all shelf objects but keep shelf itself.
}
