% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shelve.R
\name{shelf_files}
\alias{shelf_files}
\title{List files in shelf}
\usage{
shelf_files(shelf = NULL, full_names = FALSE)
}
\arguments{
\item{shelf}{Name of shelf. If `NULL, the active shelf is used.}

\item{full_names}{(bool) Full file path?}
}
\description{
List files in shelf
}
