% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shelve.R
\name{shelve}
\alias{shelve}
\title{Add to shelf}
\usage{
shelve(x, name, shelf = shelf_active())
}
\arguments{
\item{x}{Object to add to shelf.}

\item{name}{Name of obejcts to get from shelf.}

\item{shelf}{Name of shelf.}
}
\description{
Add to shelf
}
