% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shelve.R
\name{unshelve}
\alias{unshelve}
\title{Get from shelf}
\usage{
unshelve(name, shelf = shelf_active())
}
\arguments{
\item{name}{Name of obejcts to get from shelf.}

\item{shelf}{Name of shelf.}
}
\description{
Get from shelf
}
